+++ CloudShell.Hosting/ResourceManager/ResourceManagerViews.cs
@@ -6,6 +6,8 @@ public static class ResourceManagerViews
 
     public const string ResourceGraph = "cloudshell.resource-manager.resources.graph";
 
+    public const string Environment = "cloudshell.resource-manager.environment";
+
     public const string Health = "cloudshell.resource-manager.health";
 
     public const string AddResource = "cloudshell.resource-manager.resources.add";